feat: Add language selection functionality
This commit adds the functionality to select different languages in the application. It introduces a new method `showLocalesList` in the `App.vue` component that toggles the display of a list of available locales. The list is populated with the available locales from the Vue-i18n configuration. When a locale is selected, the application's language is updated accordingly. This feature enhances the user experience by allowing them to switch between different languages easily.