feat: Add API service for mental state exams
The commit adds a new API service file, `nursingApi.service.js`, which provides methods for fetching data related to mental state exams. This service will be used in the `page-mental-state-exams.vue` component to retrieve information about exams, examiners, and patients. This addition enhances the functionality of the application by enabling the retrieval of specific data for the mental state exams feature.